requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework
python-decouple

djangorestframework-simplejwt


django-filter
drf-yasg
celery
redis